doc_search/requirements.txt
certifi==2020.6.20
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.0
decorator==4.4.2
dnspython==1.16.0
etelemetry==0.2.1
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
httplib2==0.18.1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
networkx==2.4
nibabel==3.1.1
nipype==1.5.0
nltk==3.5
numpy==1.19.0
packaging==20.4
pandas==1.0.5
Pillow==7.2.0
prov==1.5.3
pydot==1.4.1
pydotplus==2.0.2
pymongo==3.10.1
PyMuPDF==1.17.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
pyxnat==1.3
rdflib==5.0.0
regex==2020.6.8
requests==2.24.0
scipy==1.5.1
simplejson==3.17.0
six==1.15.0
tqdm==4.47.0
traits==6.1.0
urllib3==1.25.9
Werkzeug==1.0.1
XlsxWriter==1.2.9
